// src/app/api/generated/model/antecedente-response.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Antecedente clinico de una historia (RF-M09-002)
 */
export interface AntecedenteResponse { 
    /**
     * Motivo declarado de la baja, o null
     */
    deactivationReason?: string;
    /**
     * Instante UTC de la baja logica, o null
     */
    deletedAt?: string;
    /**
     * Texto clinico del antecedente. No se edita: se da de baja y se registra el nuevo
     */
    descripcion?: string;
    /**
     * Identificador del antecedente
     */
    id?: number;
    /**
     * Instante UTC en que se registro
     */
    registradoEn?: string;
    /**
     * Cuenta que lo registro
     */
    registradoPor?: number;
    /**
     * Clase de antecedente
     */
    tipo?: AntecedenteResponseTipoEnum;
    /**
     * Version del registro
     */
    version?: number;
    /**
     * Si el antecedente sigue vigente
     */
    vigente?: boolean;
}
export enum AntecedenteResponseTipoEnum {
    MEDICO = 'MEDICO',
    QUIRURGICO = 'QUIRURGICO',
    ALERGIA = 'ALERGIA',
    MEDICACION = 'MEDICACION',
    FAMILIAR = 'FAMILIAR',
    HABITO = 'HABITO',
    OTRO = 'OTRO'
};